review + abstract RequestContext
