+++ clients/client/src/test/java/org/projectnessie/client/auth/TestBearerAuthenticationProvider.java
@@ -33,6 +33,7 @@ import org.projectnessie.client.http.HttpClient;
 import org.projectnessie.client.http.RequestContext;
 import org.projectnessie.client.http.RequestFilter;
 import org.projectnessie.client.http.impl.HttpHeaders;
+import org.projectnessie.client.http.impl.RequestContextImpl;
 
 @Execution(ExecutionMode.CONCURRENT)
 class TestBearerAuthenticationProvider {
@@ -94,7 +95,7 @@ class TestBearerAuthenticationProvider {
     assertThat(authFilter[0]).isInstanceOf(RequestFilter.class);
 
     HttpHeaders headers = new HttpHeaders();
-    RequestContext context = new RequestContext(headers, null, null, null);
+    RequestContext context = new RequestContextImpl(headers, null, null, null);
     authFilter[0].filter(context);
 
     assertThat(headers.asMap())
